Add Portuguese, a deck swatch, and a gate for the meaning page
Settings shows a 20x36 card from the selected deck, left of its name. The
names say which deck; only the picture says what it looks like, which is the
reason to switch. cardDrawFaceScaled() falls back to a full-size card when the
bitmap is missing, which would paint over the screen from inside a 46 px row,
so the swatch asks for the bitmap first and skips itself without one.

pt-BR carries the UI and the Rider-Waite-Smith deck. GPTarot and Marseille are
not translated yet. langApply() now falls back per pack instead of per
language, so a half-finished language shows the decks it has rather than
reverting the whole device to English, and build_strings.py reports a missing
deck instead of refusing to build.

tools/meaning_fit.py is new. uiMeaning() stops drawing after six lines and says
nothing, so an over-long meaning loses its last sentence with no error
anywhere. English RWS runs to 5 of those 6 lines. Portuguese RWS runs to 6.

Hint lines have no gate and were measured by hand. Three Portuguese hints
overflowed the chord. So did one English hint: TAP AN OPTION TO CHANGE rendered
190 px into the 160 px chord at baseline 447 and had been clipped since it was
written. It is now TAP TO CHANGE.

Portuguese needs the row names to carry no article, because every preposition
that precedes them contracts with one. The templates supply the contracted
form. That works only because all four bands are masculine singular, which the
translation had to choose deliberately.

Verified: Portuguese readings page the same as English, worst 4 pages, none
over. No meaning is over six lines. Also fixed enum arithmetic that the
compiler deprecates.

+++ tools/dump_readings.cpp
@@ -87,9 +87,9 @@ int main(int argc, char **argv) {
         printf("\"hidden\":%d,\"hidden_name\":\"%s\",", hidden,
                deckCard(text, hidden).name);
         printf("\"elements\":[\"%s\",\"%s\",\"%s\"],",
-               packGet(ui, (uint16_t)(ENG_EL_FIRE + deckCard(text, a).element)),
-               packGet(ui, (uint16_t)(ENG_EL_FIRE + deckCard(text, b).element)),
-               packGet(ui, (uint16_t)(ENG_EL_FIRE + deckCard(text, c).element)));
+               packGet(ui, (uint16_t)((uint16_t)ENG_EL_FIRE + deckCard(text, a).element)),
+               packGet(ui, (uint16_t)((uint16_t)ENG_EL_FIRE + deckCard(text, b).element)),
+               packGet(ui, (uint16_t)((uint16_t)ENG_EL_FIRE + deckCard(text, c).element)));
         printf("\"len\":%zu,\"text\":\"", len);
         jsonEscape(buf);
         printf("\"}\n");
